expose __experimentalGetMatchingVariation
