+++ packages/block-editor/src/components/block-variation-transforms/index.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { isMatch } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
@@ -16,15 +11,10 @@ import { useSelect, useDispatch } from '@wordpress/data';
 import { useState, useEffect } from '@wordpress/element';
 import { chevronDown } from '@wordpress/icons';
 
-export const getMatchingVariationName = ( blockAttributes, variations ) => {
-	if ( ! variations || ! blockAttributes ) return;
-	const matches = variations.filter( ( { attributes } ) => {
-		if ( ! attributes || ! Object.keys( attributes ).length ) return false;
-		return isMatch( blockAttributes, attributes );
-	} );
-	if ( matches.length !== 1 ) return;
-	return matches[ 0 ].name;
-};
+/**
+ * Internal dependencies
+ */
+import { __experimentalGetMatchingVariation as getMatchingVariation } from '../../utils';
 
 function __experimentalBlockVariationTransforms( { blockClientId } ) {
 	const [ selectedValue, setSelectedValue ] = useState();
@@ -46,7 +36,7 @@ function __experimentalBlockVariationTransforms( { blockClientId } ) {
 	);
 	useEffect( () => {
 		setSelectedValue(
-			getMatchingVariationName( blockAttributes, variations )
+			getMatchingVariation( blockAttributes, variations )?.name
 		);
 	}, [ blockAttributes, variations ] );
 	if ( ! variations?.length ) return null;
