+++ packages/block-editor/src/utils/index.js
@@ -1,2 +1,3 @@
 export { default as transformStyles } from './transform-styles';
 export * from './theme';
+export * from './block-variation-transforms';
